Move most of the BASIC keywords out of the lexer
This makes the lexer mostly language agnostic.  It still knows about
three BASIC specific keywords:

1. REM
2. PROC
3. FN

It would be easy enough to abstract out REM but more difficult to
abstract out FN and PROC.  I'll leave them in there for now.  The
lexer should now be generic enough to be used by the backends for
their assemblers.
