+++ frontend/basic_keywords.h
@@ -1,0 +1,170 @@
+/*
+ * Copyright (c) 2020 Mark Ryan
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#ifndef __SUBTILIS_BASIC_KEYWORD_H__
+#define __SUBTILIS_BASIC_KEYWORD_H__
+
+#include "../common/keywords.h"
+
+/* clang-format off */
+enum {
+	SUBTILIS_KEYWORD_ABS = SUBTILIS_KEYWORD_COMMON_MAX,
+	SUBTILIS_KEYWORD_ACS,
+	SUBTILIS_KEYWORD_ADVAL,
+	SUBTILIS_KEYWORD_AND,
+	SUBTILIS_KEYWORD_ASC,
+	SUBTILIS_KEYWORD_ASN,
+	SUBTILIS_KEYWORD_ATN,
+	SUBTILIS_KEYWORD_BEAT,
+	SUBTILIS_KEYWORD_BEATS,
+	SUBTILIS_KEYWORD_BGET_HASH,
+	SUBTILIS_KEYWORD_BPUT_HASH,
+	SUBTILIS_KEYWORD_BY,
+	SUBTILIS_KEYWORD_CALL,
+	SUBTILIS_KEYWORD_CASE,
+	SUBTILIS_KEYWORD_CHR_STR,
+	SUBTILIS_KEYWORD_CIRCLE,
+	SUBTILIS_KEYWORD_CLG,
+	SUBTILIS_KEYWORD_CLOSE_HASH,
+	SUBTILIS_KEYWORD_CLS,
+	SUBTILIS_KEYWORD_COLOR,
+	SUBTILIS_KEYWORD_COLOUR,
+	SUBTILIS_KEYWORD_COS,
+	SUBTILIS_KEYWORD_COUNT,
+	SUBTILIS_KEYWORD_DEF,
+	SUBTILIS_KEYWORD_DEG,
+	SUBTILIS_KEYWORD_DIM,
+	SUBTILIS_KEYWORD_DIV,
+	SUBTILIS_KEYWORD_DRAW,
+	SUBTILIS_KEYWORD_ELLIPSE,
+	SUBTILIS_KEYWORD_ELSE,
+	SUBTILIS_KEYWORD_END,
+	SUBTILIS_KEYWORD_ENDCASE,
+	SUBTILIS_KEYWORD_ENDERROR,
+	SUBTILIS_KEYWORD_ENDIF,
+	SUBTILIS_KEYWORD_ENDPROC,
+	SUBTILIS_KEYWORD_ENDWHILE,
+	SUBTILIS_KEYWORD_EOF_HASH,
+	SUBTILIS_KEYWORD_EOR,
+	SUBTILIS_KEYWORD_ERL,
+	SUBTILIS_KEYWORD_ERR,
+	SUBTILIS_KEYWORD_ERROR,
+	SUBTILIS_KEYWORD_EVAL,
+	SUBTILIS_KEYWORD_EXP,
+	SUBTILIS_KEYWORD_EXT_HASH,
+	SUBTILIS_KEYWORD_FALSE,
+	SUBTILIS_KEYWORD_FILL,
+	SUBTILIS_KEYWORD_FOR,
+	SUBTILIS_KEYWORD_GCOL,
+	SUBTILIS_KEYWORD_GET,
+	SUBTILIS_KEYWORD_GET_STR,
+	SUBTILIS_KEYWORD_GET_STR_HASH,
+	SUBTILIS_KEYWORD_HEAP_FREE,
+	SUBTILIS_KEYWORD_IF,
+	SUBTILIS_KEYWORD_INKEY,
+	SUBTILIS_KEYWORD_INKEY_STR,
+	SUBTILIS_KEYWORD_INPUT,
+	SUBTILIS_KEYWORD_INPUT_HASH,
+	SUBTILIS_KEYWORD_INSTR,
+	SUBTILIS_KEYWORD_INT,
+	SUBTILIS_KEYWORD_LEFT_STR,
+	SUBTILIS_KEYWORD_LEN,
+	SUBTILIS_KEYWORD_LET,
+	SUBTILIS_KEYWORD_LINE,
+	SUBTILIS_KEYWORD_LN,
+	SUBTILIS_KEYWORD_LOCAL,
+	SUBTILIS_KEYWORD_LOG,
+	SUBTILIS_KEYWORD_MID_STR,
+	SUBTILIS_KEYWORD_MOD,
+	SUBTILIS_KEYWORD_MODE,
+	SUBTILIS_KEYWORD_MOUSE,
+	SUBTILIS_KEYWORD_MOVE,
+	SUBTILIS_KEYWORD_NEXT,
+	SUBTILIS_KEYWORD_NOT,
+	SUBTILIS_KEYWORD_OF,
+	SUBTILIS_KEYWORD_OFF,
+	SUBTILIS_KEYWORD_ON,
+	SUBTILIS_KEYWORD_ONERROR,
+	SUBTILIS_KEYWORD_OPENIN,
+	SUBTILIS_KEYWORD_OPENOUT,
+	SUBTILIS_KEYWORD_OPENUP,
+	SUBTILIS_KEYWORD_OR,
+	SUBTILIS_KEYWORD_ORIGIN,
+	SUBTILIS_KEYWORD_OSCLI,
+	SUBTILIS_KEYWORD_OTHERWISE,
+	SUBTILIS_KEYWORD_PI,
+	SUBTILIS_KEYWORD_PLOT,
+	SUBTILIS_KEYWORD_POINT,
+	SUBTILIS_KEYWORD_POS,
+	SUBTILIS_KEYWORD_PRINT,
+	SUBTILIS_KEYWORD_PRINT_HASH,
+	SUBTILIS_KEYWORD_PTR_HASH,
+	SUBTILIS_KEYWORD_QUIT,
+	SUBTILIS_KEYWORD_RAD,
+	SUBTILIS_KEYWORD_RECTANGLE,
+	SUBTILIS_KEYWORD_REPEAT,
+	SUBTILIS_KEYWORD_REPORT,
+	SUBTILIS_KEYWORD_REPORT_STR,
+	SUBTILIS_KEYWORD_RETURN,
+	SUBTILIS_KEYWORD_RIGHT_STR,
+	SUBTILIS_KEYWORD_RND,
+	SUBTILIS_KEYWORD_SGN,
+	SUBTILIS_KEYWORD_SIN,
+	SUBTILIS_KEYWORD_SOUND,
+	SUBTILIS_KEYWORD_SPC,
+	SUBTILIS_KEYWORD_SQR,
+	SUBTILIS_KEYWORD_STEP,
+	SUBTILIS_KEYWORD_STEREO,
+	SUBTILIS_KEYWORD_STOP,
+	SUBTILIS_KEYWORD_STRING_STR,
+	SUBTILIS_KEYWORD_STR_STR,
+	SUBTILIS_KEYWORD_SUM,
+	SUBTILIS_KEYWORD_SUMLEN,
+	SUBTILIS_KEYWORD_SWAP,
+	SUBTILIS_KEYWORD_SYS,
+	SUBTILIS_KEYWORD_TAB,
+	SUBTILIS_KEYWORD_TAN,
+	SUBTILIS_KEYWORD_TEMPO,
+	SUBTILIS_KEYWORD_THEN,
+	SUBTILIS_KEYWORD_TIME,
+	SUBTILIS_KEYWORD_TIME_STR,
+	SUBTILIS_KEYWORD_TINT,
+	SUBTILIS_KEYWORD_TO,
+	SUBTILIS_KEYWORD_TRUE,
+	SUBTILIS_KEYWORD_UNTIL,
+	SUBTILIS_KEYWORD_USR,
+	SUBTILIS_KEYWORD_VAL,
+	SUBTILIS_KEYWORD_VDU,
+	SUBTILIS_KEYWORD_VOICES,
+	SUBTILIS_KEYWORD_VPOS,
+	SUBTILIS_KEYWORD_WAIT,
+	SUBTILIS_KEYWORD_WHEN,
+	SUBTILIS_KEYWORD_WHILE,
+	SUBTILIS_KEYWORD_WIDTH,
+	SUBTILIS_KEYWORD_MAX
+};
+
+/* clang-format on */
+
+/*
+ * PROC and FN are upper case only.
+ */
+
+#define SUBTILIS_KEYWORD_TOKENS ((SUBTILIS_KEYWORD_MAX * 2) - 2)
+
+extern const subtilis_keyword_t subtilis_keywords_list[SUBTILIS_KEYWORD_TOKENS];
+
+#endif
