+++ frontend/parser_compound.h
@@ -23,8 +23,7 @@
 void subtilis_parser_local(subtilis_parser_t *p, subtilis_token_t *t,
 			   subtilis_error_t *err);
 void subtilis_parser_compound(subtilis_parser_t *p, subtilis_token_t *t,
-			      subtilis_keyword_type_t end_key,
-			      subtilis_error_t *err);
+			      int end_key, subtilis_error_t *err);
 void subtilis_parser_statement(subtilis_parser_t *p, subtilis_token_t *t,
 			       subtilis_error_t *err);
